+++ src/Cloud/DTO/DatabaseDependencyType.php
@@ -10,4 +10,7 @@ enum DatabaseDependencyType: string
     case OWNED_DATABASE_CHILD = 'owned_database_child';
     case UNMANAGED_DATABASE_CHILD = 'unmanaged_database_child';
     case OWNERSHIP_CONFLICT = 'ownership_conflict';
+    case DATABASE_SNAPSHOT = 'database_snapshot';
+    case RETAINED_DATABASE_RECOVERY = 'retained_database_recovery';
+    case DATABASE_CLUSTER_LIFECYCLE = 'database_cluster_lifecycle';
 }